requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tkinterdnd2==0.4.2
tzdata==2025.1
